user authentication set up

+++ task-manager/app/Http/Controllers/TaskListController.php
@@ -4,17 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Models\TaskList;
 use Illuminate\Http\Request;
+use App\Models\User;
+use App\Http\Controllers\Auth;
 
 // THIS IS FOR THE LIST
 class TaskListController extends Controller
 {
-    public function index(){
-        $lists = TaskList::all();
-        return response()->json($lists, 200);
+    public function index(User $user){
+        return response()->json($user->lists, 200);
     }
 
-    public function show ($id){
-        $list= TaskList::find($id);
+    public function show (TaskList $list){
+        if (Auth::id() !== $list->user_id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
+
         return response()->json($list, 200);
     }
 
@@ -22,14 +26,17 @@ class TaskListController extends Controller
     public function store(){
         $list = new TaskList();
         $list->list_name = request('list_name');
+        $list->user_id = Auth::id();
         $list->save(); //save to database
     
         return response()->json($list, 201);
     }
     
     // update
-    public function update($id){
-        $list = TaskList::find($id);
+    public function update(Request $request, TaskList $list){
+        if (Auth::id() !== $list->user_id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         $list->list_name = request('list_name');
         $list->save(); //save to database
     
@@ -37,8 +44,10 @@ class TaskListController extends Controller
     }
     
     // delete list
-    public function destroy($id){
-        $list = TaskList::find($id);
+    public function destroy(TaskList $list){
+        if (Auth::id() !== $list->user_id) {
+            return response()->json(['error' => 'Unauthorized'], 403);
+        }
         $list->delete();
 
         return response()->json($list, 200);
